pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["gledsonfagno <gleidsonfagner838@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.1"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
